crawler/requirements.txt
beautifulsoup4==4.12.3
crawler==0.0.2
Flask==3.0.2
python_crontab==3.0.0
Requests==2.31.0
whoosh==2.7.4
